Refactor code. Fix gluing operation
