+++ rule generator/RuleGenerator/RuleGenerator.vcxproj
@@ -167,7 +167,6 @@
     <ClInclude Include="..\..\cpp_version\graph\graph_half_edge.h" />
     <ClInclude Include="..\..\cpp_version\graph\graph_vertex.h" />
     <ClInclude Include="..\..\cpp_version\memory_counter.h" />
-    <ClInclude Include="MatchGraph.h" />
     <ClInclude Include="..\..\cpp_version\primitives\edge_type.h" />
     <ClInclude Include="..\..\cpp_version\primitives\face_type.h" />
     <ClInclude Include="..\..\cpp_version\primitives\primitives.h" />
@@ -224,7 +223,6 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="MatchGraph.cpp" />
     <ClCompile Include="..\..\cpp_version\primitives\edge_type.cpp" />
     <ClCompile Include="..\..\cpp_version\primitives\face_type.cpp" />
     <ClCompile Include="..\..\cpp_version\primitives\primitives.cpp" />
